requirements.txt
fastapi
uvicorn
python-multipart
pydantic>=2.0
requests
numpy
opencv-python-headless
pdf2image
google-generativeai
pillow
python-dotenv